Committin

+++ database/migrations/2016_03_27_222827_create_contacts_table.php
@@ -14,9 +14,17 @@ class CreateContactsTable extends Migration
     {
         Schema::create('contacts', function (Blueprint $table) {
             $table->increments('id');
+            $table->integer('user_id')->unsigned();
+            $table->integer('image_id')->unsigned()->nullable();
+            $table->string('url_name')->unique();
             $table->string('firstname');
+            $table->string('middlename');
             $table->string('lastname');
-            $table->boolean('some_bool');
+            $table->string('phone');
+            $table->string('email');
+            $table->string('address');
+            $table->integer('age');
+            $table->date('birthday');
             $table->timestamps();
         });
     }
@@ -28,6 +36,9 @@ class CreateContactsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('contacts');
+        if (Schema::hasTable('contacts'))
+        {
+            Schema::drop('contacts');
+        }
     }
 }
